Name all threads, some fixes, add ServerCrashAutoRestart
ServerCrashAutoRestart is a test feature
it will try to restart your server if it will crash or main thread will
freeze(freeze- probably not always).
Can't guarantee that it will work
Also if you stop server using stop command from Chat, there is a chance
that AutoSaveWorld will think that this is a crash and restart server.
/asw test command will restart server
if you will use spigot /restart command and it will not restart server,
ServerCrashAutoRestart will stop functioning.